+++ src/util/compendium-util.ts
@@ -5,10 +5,17 @@ export default class CompendiumUtil {
   private static compendiumCareerGroups: string[];
   private static compendiumTrappings: Item[];
   private static compendiumBestiary: { [pack: string]: Actor[] };
+  private static compendiumSkills: Item[];
+  private static compendiumTalents: Item[];
+  private static compendiumTraits: Item[];
+  private static compendiumSizeTrait: Item;
+  private static compendiumSwarmTrait: Item;
+  private static compendiumWeaponTrait: Item;
+  private static compendiumArmorTrait: Item;
 
   private static compendiumLoaded = false;
 
-  public static async initCompendium(withCreatures = false) {
+  public static async initCompendium(forCreatures = false) {
     let loadDialog: Dialog | null = null;
     if (!this.compendiumLoaded) {
       loadDialog = new Dialog({
@@ -23,11 +30,20 @@ export default class CompendiumUtil {
       });
       loadDialog.render(true);
     }
-    await this.getCompendiumCareers();
-    await this.getCompendiumCareersGroups();
+    if (!forCreatures) {
+      await this.getCompendiumCareers();
+      await this.getCompendiumCareersGroups();
+    }
     await this.getCompendiumTrappings();
-    if (withCreatures) {
+    if (forCreatures) {
       await this.getCompendiumBestiary();
+      await this.getCompendiumSkills();
+      await this.getCompendiumTalents();
+      await this.getCompendiumTraits();
+      await this.getCompendiumSizeTrait();
+      await this.getCompendiumSwarmTrait();
+      await this.getCompendiumWeaponTrait();
+      await this.getCompendiumArmorTrait();
     }
     if (!this.compendiumLoaded && loadDialog != null) {
       setTimeout(async () => {
@@ -41,7 +57,9 @@ export default class CompendiumUtil {
     if (this.compendiumCareers == null) {
       this.compendiumCareers = [];
       const careersPacks = game.packs.filter(
-        (p) => p.metadata.tags && p.metadata.tags.includes('career')
+        (p) =>
+          (p.metadata.tags && p.metadata.tags.includes('career')) ||
+          p.metadata.entity === 'Item'
       );
       for (let pack of careersPacks) {
         const career: Item[] = await pack.getContent();
@@ -75,7 +93,9 @@ export default class CompendiumUtil {
     if (this.compendiumTrappings == null) {
       this.compendiumTrappings = [];
       const trappingsPacks = game.packs.filter(
-        (p) => p.metadata.tags && p.metadata.tags.includes('trapping')
+        (p) =>
+          (p.metadata.tags && p.metadata.tags.includes('trapping')) ||
+          p.metadata.entity === 'Item'
       );
       const trappingCategories = CompendiumUtil.getTrappingCategories();
       for (let pack of trappingsPacks) {
@@ -112,5 +132,99 @@ export default class CompendiumUtil {
       }
     }
     return Promise.resolve(this.compendiumBestiary);
+  }
+
+  public static async getCompendiumSkills() {
+    if (this.compendiumSkills == null) {
+      this.compendiumSkills = [];
+      const skillsPacks = game.packs.filter(
+        (p) =>
+          (p.metadata.tags && p.metadata.tags.includes('skill')) ||
+          p.metadata.entity === 'Item'
+      );
+      for (let pack of skillsPacks) {
+        const career: Item[] = await pack.getContent();
+        this.compendiumSkills.push(...career.filter((c) => c.type === 'skill'));
+      }
+    }
+    return Promise.resolve(this.compendiumSkills);
+  }
+
+  public static async getCompendiumTalents() {
+    if (this.compendiumTalents == null) {
+      this.compendiumTalents = [];
+      const talentsPacks = game.packs.filter(
+        (p) =>
+          (p.metadata.tags && p.metadata.tags.includes('talent')) ||
+          p.metadata.entity === 'Item'
+      );
+      for (let pack of talentsPacks) {
+        const career: Item[] = await pack.getContent();
+        this.compendiumTalents.push(
+          ...career.filter((c) => c.type === 'talent')
+        );
+      }
+    }
+    return Promise.resolve(this.compendiumTalents);
+  }
+
+  public static async getCompendiumTraits() {
+    if (this.compendiumTraits == null) {
+      this.compendiumTraits = [];
+      const traitsPacks = game.packs.filter(
+        (p) =>
+          (p.metadata.tags && p.metadata.tags.includes('trait')) ||
+          p.metadata.entity === 'Item'
+      );
+      for (let pack of traitsPacks) {
+        const career: Item[] = await pack.getContent();
+        this.compendiumTraits.push(...career.filter((c) => c.type === 'trait'));
+      }
+    }
+    return Promise.resolve(this.compendiumTraits);
+  }
+
+  public static async getCompendiumSizeTrait() {
+    if (this.compendiumSizeTrait == null) {
+      this.compendiumSizeTrait = <Item>(
+        (await this.getCompendiumTraits()).find(
+          (t: Item & any) => t.name === 'Size' || t.originalName === 'Size'
+        )
+      );
+    }
+    return Promise.resolve(this.compendiumSizeTrait);
+  }
+
+  public static async getCompendiumSwarmTrait() {
+    if (this.compendiumSwarmTrait == null) {
+      this.compendiumSwarmTrait = <Item>(
+        (await this.getCompendiumTraits()).find(
+          (t: Item & any) => t.name === 'Swarm' || t.originalName === 'Swarm'
+        )
+      );
+    }
+    return Promise.resolve(this.compendiumSwarmTrait);
+  }
+
+  public static async getCompendiumWeaponTrait() {
+    if (this.compendiumWeaponTrait == null) {
+      this.compendiumWeaponTrait = <Item>(
+        (await this.getCompendiumTraits()).find(
+          (t: Item & any) => t.name === 'Weapon' || t.originalName === 'Weapon'
+        )
+      );
+    }
+    return Promise.resolve(this.compendiumWeaponTrait);
+  }
+
+  public static async getCompendiumArmorTrait() {
+    if (this.compendiumArmorTrait == null) {
+      this.compendiumArmorTrait = <Item>(
+        (await this.getCompendiumTraits()).find(
+          (t: Item & any) => t.name === 'Armour' || t.originalName === 'Armour'
+        )
+      );
+    }
+    return Promise.resolve(this.compendiumArmorTrait);
   }
 }