+++ src/util/creature-abilities-chooser.ts
@@ -1,0 +1,7 @@
+export default class CreatureAbilitiesChooser {
+  // public static async selectCreatureAbilities(
+  //     callback: (abilities) => void
+  // ) {
+  //
+  // }
+}